+++ lib/rails_age.rb
@@ -11,7 +11,7 @@ module ApacheAge
   require 'apache_age/entities/edge'
   require 'apache_age/entities/entity'
   require 'apache_age/entities/vertex'
+  require 'apache_age/validators/unique_edge'
+  require 'apache_age/validators/unique_vertex'
   require 'apache_age/types/age_type_generator'
-  require 'apache_age/validators/unique_edge_validator'
-  require 'apache_age/validators/unique_vertex_validator'
 end