+++ spec/dummy/spec/lib/generators/apache_age/node_generator_spec.rb
@@ -9,23 +9,23 @@ RSpec.describe ApacheAge::NodeGenerator, type: :generator do
   let(:config) { {behavior:, destination_root:} }
   let(:args) { [node_name, "name", "age:integer"] }
   let(:expected_node_content) {
-  <<~NODEFILE
-  class #{node_name.split('/').join('::')}
-    include ApacheAge::Entities::Vertex
+    <<~NODE_FILE
+    class #{node_name.split('/').join('::')}
+      include ApacheAge::Entities::Vertex
 
-    attribute :name, :string
-    attribute :age, :integer
+      attribute :name, :string
+      attribute :age, :integer
 
-    validates :name, presence: true
-    validates :age, presence: true
+      validates :name, presence: true
+      validates :age, presence: true
 
-    # unique node validator (remove any attributes that are not important to uniqueness)
-    validates_with(
-      ApacheAge::Validators::UniqueVertexValidator,
-      attributes: [:name, :age]
-    )
-  end
-  NODEFILE
+      # custom unique node validator (remove any attributes that are NOT important to uniqueness)
+      validates_with(
+        ApacheAge::Validators::UniqueVertex,
+        attributes: [:name, :age]
+      )
+    end
+    NODE_FILE
   }
 
   before { described_class.start([node_name], {behavior: :revoke, destination_root:}) }
@@ -61,7 +61,7 @@ RSpec.describe ApacheAge::NodeGenerator, type: :generator do
     end
 
     context 'with a namespace' do
-      let(:node_name) { "Test/TestNode"}
+      let(:node_name) { "Tests/TestNode"}
 
       it "inserts correct name and date" do
         described_class.start(args, config)
@@ -72,18 +72,22 @@ RSpec.describe ApacheAge::NodeGenerator, type: :generator do
         expect(file_content).to eq(expected_node_content)
         type_content = File.read(types_config_path)
         expect(type_content)
-          .to include("require_dependency 'test/test_node'")
+          .to include("require_dependency 'tests/test_node'")
         expect(type_content)
-          .to include(":test_test_node, ApacheAge::Types::AgeTypeGenerator.create_type_for(Test::TestNode)")
+          .to include(
+            ":tests_test_node, ApacheAge::Types::AgeTypeGenerator.create_type_for(Tests::TestNode)"
+          )
 
         # ensure destroy works
         described_class.start([node_name], {behavior: :revoke, destination_root:})
         expect(File.exist?(expected_node_path)).to be false
         type_content = File.read(types_config_path)
         expect(type_content)
-          .not_to include("require_dependency 'test/test_node'")
+          .not_to include("require_dependency 'tests/test_node'")
         expect(type_content)
-          .not_to include(":test_test_node, ApacheAge::Types::AgeTypeGenerator.create_type_for(Test::TestNode)")
+          .not_to include(
+            ":tests_test_node, ApacheAge::Types::AgeTypeGenerator.create_type_for(Tests::TestNode)"
+          )
       end
     end
   end