+++ CHANGELOG.md
@@ -23,17 +23,28 @@
 - **Edge Scaffold** (generates edge, type, view and controller)
   * add `rails generate apache_age:edge_scaffold HasJob employee_role start_node:person end_node:company`
 
-## VERSION 0.4.2 - 2024-xx-xx
+## VERSION 0.5.2 - 2024-xx-xx
 
 - **Node Scaffold** (generates node, type, view and controller)
   * add `rails generate apache_age:node_scaffold Person first_name last_name age:integer`
 
-## VERSION 0.4.2 - 2024-xx-xx
+## VERSION 0.5.1 - 2024-xx-xx
 
-- **Edge Generator**
-  * add `rails generate apache_age:edge HasPet owner_role` just a property
+- **Edge Generator** (hopefully - I can figure out how to load custom types into the generators
+  * add `rails generate apache_age:edge HasPet owner_role start_node:person end_node:pet`
+        add edge type generation
   * add `rails generate apache_age:edge HasPet owner_role start_node:person end_node:pet`
         with property and specified start-/end-nodes (person and pet nodes must have already been created)
+
+## VERSION 0.5.0 - 2024-06-16
+
+**breaking change**: renamed the validators to UniqueVertex and UniqueEdge
+
+- **Edge Generator**
+  * add `rails generate apache_age:edge HasPet owner_role`
+    caveats:
+    1. start_node and end_node are of type `:vertex` in the generator but can be changed manually in the class file - having trouble with the generator loading the types (so it the generator rejects custom type - but rails still works with them)
+    2. edge types are not yet generated automatically (not yet needed)
 
 ## VERSION 0.4.1 - 2024-06-15
 
@@ -43,7 +54,7 @@
 
 ## VERSION 0.4.0 - 2024-06-14
 
-Minor breaking change: type (:vertix) is now required in core for edges
+**breaking change**: type (:vertix) is now required in core for edges
 
 - **Installer**
   * AGE types added to installer (with tests)